requirements.txt
flask>=3.1.1
elasticsearch>=9.0.3
python-dotenv>=1.0.1
requests>=2.31.0
openai>=1.54.1
numpy>=1.26.4
pandas>=2.2.1 